+++ CHANGELOG.md
@@ -7,11 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.1] - 2022-02-11
+
+### Fixed
+
+- Redirect from '/jupyterlite' to '/jupyterlite/' since trying to load '/jupyterlite' doesn't work
+- Misc cleanup in the top-level controller
+
 ## [1.0.0] - 2022-02-11
 
 ### Added
 
 - Initial release
 
-[unreleased]: https://github.com/symbioquine/drupal_jupyterlite/compare/unbuilt-v1.0.0...HEAD
+[unreleased]: https://github.com/symbioquine/drupal_jupyterlite/compare/unbuilt-v1.0.1...HEAD
+[1.0.1]: https://github.com/symbioquine/drupal_jupyterlite/compare/unbuilt-v1.0.0...unbuilt-v1.0.1
 [1.0.0]: https://github.com/symbioquine/drupal_jupyterlite/releases/tag/unbuilt-v1.0.0